#55 Refactor head and tail counters to atomics
Head and Tail on ChannelRaw get accessed in channelSend and channelReceive
before a lock acquisition happens in those procs.
That means that the output of the template they get used in (isEmpty/isFull) is dependent on a data-race.

That is because as a thread that has the lock may modify those values while they are being
read by another thrread without the lock.
This is noticed by thread-sanitizer and should be eliminated for the following reasons:
- Users should not catch "false positives" for thread sanitizier such as this, which they currently will
- It is generally a good idea to have multi-threaded code such as this data-race free

This refactor to atomics makes the entire thing atomic and thus impossible to have a
data-race with.